Moves fetchMessages to state-store.js
